backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "makeme-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.6.0",
    "PyJWT>=2.10.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "cryptography>=44.0.0",
    "numpy>=2.0.0",
    "icalendar>=6.0.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "pytest-cov>=6.0.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app", "workers"]
